+++ cmd/foreverdubbed/desktop_stub.go
@@ -7,7 +7,7 @@ import (
 	"foreverdubbed/internal/appstate"
 )
 
-func runDesktop(_ context.Context, _ context.CancelFunc, _ *appstate.State, work func() error) error {
+func runDesktop(_ context.Context, _ context.CancelFunc, _ *appstate.State, _ []string, work func() error) error {
 	return work()
 }